Removes Python 2.7 support (#2124)
* Adds pycodestyle for PyCharm PEP8 analysis

* Removal of wxPython version conditions, code cleanup phase 1

* Removal of wxPython version conditions, code cleanup phase 1

* Removal of wxPython version conditions, code cleanup phase 1

* Removal of wxPython version conditions, code cleanup phase 1

* Removal of wxPython version conditions, code cleanup phase 1

* Removal of wxPython version conditions, code cleanup phase 1

* Change max line length to 90 char, for PyCharm

* Main removal of PY2 and PY3

* Final removal of PY2 and PY3 and basestring, unicode and unichar functions.
