+++ src/robotide/controller/cellinfo.py
@@ -15,9 +15,6 @@
 
 from robotide.action.shortcut import localize_shortcuts
 from robotide.utils import highlightmatcher, html_escape
-from robotide.utils import PY3
-if PY3:
-    from robotide.utils import unicode
 
 CTRL_LABEL = localize_shortcuts('CtrlCmd')
 
@@ -67,7 +64,7 @@ def TipMessage(cell):
         return ''
     tip = _TooltipMessage(cell) if not cell.for_loop \
         else _ForLoopTooltipMessage(cell)
-    return html_escape(unicode(tip)).replace('\n', '<br />')
+    return html_escape(str(tip)).replace('\n', '<br />')
 
 
 class _TooltipMessage(object):
