+++ src/robotide/editor/clipboard.py
@@ -16,7 +16,6 @@
 import os
 import wx
 from robotide.context import IS_WINDOWS
-from robotide.utils import basestring
 
 
 class _ClipboardHandler(object):
@@ -132,7 +131,7 @@ class _GridClipboard(object):
     def _format_data(self, data):
         if isinstance(data, list):
             return os.linesep.join('\t'.join(row) for row in data)
-        if isinstance(data, basestring):
+        if isinstance(data, str):
             return data
         return None
