+++ utest/controller/test_occurrences.py
@@ -15,7 +15,6 @@ from robotide.publish.messages import RideItemStepsChanged,\
 from robotide.namespace.namespace import Namespace
 from robotide.spec.librarymanager import LibraryManager
 from robotide.usages.commands import FindUsages
-from robotide.utils import PY2
 from resources import FakeSettings
 import datafilereader
 
@@ -107,10 +106,9 @@ def _first_occurrence(test_ctrl, kw_name):
     occurrences = test_ctrl.execute(FindOccurrences(kw_name))
     if not occurrences:
         raise AssertionError('No occurrences found for "%s"' % kw_name)
-    if PY2:
-        return occurrences.next()  # Python 2.7
-    return next(occurrences)  # DEBUG .next() Python 3
-    # see https://stackoverflow.com/questions/21622193/python-3-2-coroutine-attributeerror-generator-object-has-no-attribute-next
+    return next(occurrences)
+    # see https://stackoverflow.com/questions/21622193/
+    # python-3-2-coroutine-attributeerror-generator-object-has-no-attribute-next
 
 
 def _get_ctrl_by_name(self, name, datafiles):
@@ -153,14 +151,8 @@ class TestFindOccurrencesWithFiles(unittest.TestCase):
 
     def test_first_occurrences_are_from_the_same_file(self):
         occ = self.resu.execute(FindOccurrences('My Keyword'))
-        if PY2:
-            # Python 2.7
-            assert_true(self.resu.filename.endswith(occ.next().item.parent.source))
-            assert_equal(occ.next().source, self.ts2.source)
-            assert_equal(occ.next().source, self.ts2.source)
-            return
-        # Python 3
-        # see https://stackoverflow.com/questions/21622193/python-3-2-coroutine-attributeerror-generator-object-has-no-attribute-next
+        # see https://stackoverflow.com/questions/21622193/
+        # python-3-2-coroutine-attributeerror-generator-object-has-no-attribute-next
         assert_true(self.resu.filename.endswith(occ.__next__().item.parent.source))
         assert_equal(occ.__next__().source, self.ts2.source)
         assert_equal(occ.__next__().source, self.ts2.source)
