+++ utest/spec/test_library_database.py
@@ -19,9 +19,6 @@ import unittest
 from robotide.spec.iteminfo import LibraryKeywordInfo
 from robotide.spec.librarydatabase import LibraryDatabase
 from robotide.spec.libraryfetcher import get_import_result
-from robotide.utils import PY3
-if PY3:
-    from robotide.utils import unicode
 
 testlibpath = os.path.join(os.path.dirname(__file__), '..', 'resources',
                            'robotdata', 'lib_with_doc_format')
@@ -92,7 +89,7 @@ class TestLibraryDatabase(unittest.TestCase):
         self._database.insert_library_keywords('lib.py', 'foo', [LibraryKeywordInfo('this is old', 'doc', 'ROBOT', 'lib.py', '')])
         self._database.insert_library_keywords('lib.py', 'foo', [LibraryKeywordInfo('this is new', 'doc', 'ROBOT', 'lib.py', '')])
         kws = self._database.fetch_library_keywords('lib.py', 'foo')
-        self.assertEqual(len(kws), 1, unicode(kws))
+        self.assertEqual(len(kws), 1, str(kws))
         self.assertEqual(kws[0].name, 'this is new')
 
     def test_removing_old_data(self):
